# Conflicts:
#	src/test/java/com/endava/InternetBankingTest.java

+++ src/main/java/com/endava/controllers/BankingController.java
@@ -7,6 +7,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
+
 @RestController
 @RequestMapping("/banking")
 public class BankingController {
@@ -17,8 +19,11 @@ public class BankingController {
     @Autowired
     private UserService userService;
 
+    // TODO: top-up account -> account
     @PutMapping("/topup/{id}")
-    public ResponseEntity topUp(@PathVariable Long id, @RequestParam Double funds) {
+    public ResponseEntity topUp(
+            @PathVariable Long id,
+            @RequestParam BigDecimal funds) {
         return userService.findById(id)
                 .filter(user -> bankingService.topUp(user.getId(), funds))
                 .map(user -> new ResponseEntity<>(user, HttpStatus.OK))
